mylinebot/requirements.txt
Django
requests
line-bot-sdk
gunicorn
emoji